requirements.txt
PyQt5==5.10
PyQt5-sip==4.19.19
PyQtChart==5.10
coverage==5.0.3
pyaml==20.4.0
